Remove driver wrapper.
